requirements.txt
Django==4.2.14
python-dotenv==1.0.1
djangorestframework==3.15.2
mysqlclient==2.2.4
